impl table

+++ src/skip_list/list.rs
@@ -18,7 +18,7 @@ pub struct Node {
     pub key: Bytes,
     value: Bytes,
     height: usize,
-    tower: [AtomicU32; MAX_HEIGHT as usize],
+    tower: [AtomicU32; MAX_HEIGHT],
 }
 
 impl Node {
@@ -26,7 +26,7 @@ impl Node {
         let align = mem::align_of::<Node>();
         let size = mem::size_of::<Node>();
         // Not all values in Node::tower will be utilized.
-        let not_used = (MAX_HEIGHT as usize - height as usize - 1) * mem::size_of::<AtomicU32>();
+        let not_used = (MAX_HEIGHT - height - 1) * mem::size_of::<AtomicU32>();
         let node_offset = arena.alloc(align, size - not_used);
         unsafe {
             let node_ptr: *mut Node = arena.get_mut(node_offset);
